pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cellseg-challenge"
version = "0.1.0"
description = "Lightning and Hydra setup for the TUM AI E-Lab cell segmentation Kaggle challenge."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "albumentations>=1.4.0",
  "hydra-core>=1.3.2",
  "lightning>=2.2.0",
  "numpy>=1.24",
  "opencv-python-headless>=4.8",
  "pillow>=10.0",
  "scipy>=1.10",
  "segmentation-models-pytorch>=0.3.3",
  "torch>=2.1",
  "torchvision>=0.16",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

